requirements.txt
annotated-types==0.7.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
openfoodfacts==3.3.0
pip==25.3
prompt_toolkit==3.0.52
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyYAML==6.0.3
questionary==2.1.1
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.2
wcwidth==0.2.14
wheel==0.45.1
